requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.18
jinja2>=3.1.0
aiofiles>=24.1.0
pydantic-settings>=2.6.0

llama-index-core>=0.12.0
llama-index-llms-openrouter>=0.3.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-readers-file>=0.4.0

chromadb>=0.6.0
pandas>=2.2.0
openpyxl>=3.1.0
xlrd>=2.0.1
pdfplumber>=0.11.0
